move PlotlySurfaceFigure tests to test_figures
